server/requirements.txt
uvicorn
fastapi
python-multipart
Pillow
tensorflow
numpy
scikit-learn